feat(eticket): conditional rendering by payment status
